修改P_TASK_DETAIL字段TPL_DETAIL_ID字段变为TASK_ID

+++ src/main/java/org/fms/eis/webapp/domain/PTaskDomain.java
@@ -21,7 +21,7 @@ public class PTaskDomain extends Page implements MybatisEntity {
     private String name;    //任务名称
     private Long protocolId;    //规约编码
     private Long taskId;    //任务类型
-    private Long taskNo;    //任务号
+    private Integer taskNo;    //任务号
     private String pnType;    //信息点类型
     private String objProp;    //对象属性
     private Date taskBegintime;    //任务起始时间
@@ -81,11 +81,11 @@ public class PTaskDomain extends Page implements MybatisEntity {
         this.taskId = taskId;
     }
 
-    public Long getTaskNo() {
+    public Integer getTaskNo() {
         return taskNo;
     }
 
-    public void setTaskNo(Long taskNo) {
+    public void setTaskNo(Integer taskNo) {
         this.taskNo = taskNo;
     }
